requirements.txt
streamlit==1.29.0
playwright==1.40.0
validators==0.22.0
python-dotenv==1.0.0
google-generativeai==0.3.1
nest-asyncio==1.6.0